Added CSS, JS, and edited the base, home, and dashboard template (#24)
* Update base.css

Separated the base.css from other CSS pages in case we wanted to specifically edit pother pages. The base.css is now a generic page.

* created a dashboard.css page

Created a dashboard.css page that is specifically for the dashboard page. The background is semi-transparent to allow the text to be seen. This page includes a background image titled "dashboard_backgorund.jpg" that will be in the img folder.

* Created a home.css file

This is the home.css file that is now working independently of base.css. It includes centered text and a picture of coins with our names underneath. There is also now dollar signs falling from the top of the screen.

* created a make-it-rain css

Created a make-it-rain css to detail the dollar signs that fall from the top of the homepage.

* dashboard_background.jpg

Here is the image used for the dashboard background. The resolution is 1920 x 1170 for reference

* created make-it-rain.js

Here is the make-it-rain.js file used to move the make-it-rain.css.

* Update base.html

Updated the file to include the code blocks for the added 
css:

    <!-- include the extra_head block for make it rain css/calculator_background -->
    {% block extra_head %}{% endblock %}

    <link rel="stylesheet" href="{% static 'css/base.css' %}">

and javascript: 

    <!-- this makes it so only the other pages can use specific css features -->
    <main>
        {% block content %}{% endblock %}
    </main>

    <!-- add block for make it rain js -->
    {% block extra_js %}{% endblock %}

* Update dashboard.html

Updated dashboard.html to include background image and css features. Created a "<body class="dashboard-main">" and "<div class="dashboard">" to structure the additions. 

Also added emojis to for razzle dazzle.

* Update home.html

Updated the home.html to include added CSS and make-it-rain.js.
Added for better structure.:
<main class="home-main"> and <div id="make-it-rain"></div>  

Added block for JavaScript: 
{% block extra_js %}
<script src="{% static 'js/make-it-rain.js' %}"></script>
{% endblock %}

* Added coins.png

Added the coins.png to reflect the image on the homepage in the proposed demo.

* Redo migrations, delete make it rain js, remove unnecessary code from home.html and base.html.

* Update home.html

Removed the "make-it-rain" code

* Update base.html

Removed the "make-it-rain" code

* Delete static/css/make-it-rain.css

Removed the "make-it-rain" code

---------

Co-authored-by: Kenneth Stancil <githubbeded@proton.me>