Edit Category
I commented out the goal object in the class Category in the models.py for this to work. I'm going to have to figure out why the goals is messing up the category class. While I was working I kept running into issues saying the database doesn't have a goal section as an object. Is there a way to add "goal" to the database?