Create time-period-sidebar
Based on Trello board item: "As a user, I want to view my income and expenses in certain time periods to figure out how much I spent or earned."

I went ahead and created another branch off master to build the time-period-sidebar. This was on the trello backlog for Sprint 3. It's not currently working because main isn't up to date with the edit_transaction branch, therefore the transactions aren't being saved. However, the basic framework is there for us to build on once everything is merged. I added to the dashboard view, updated dashboard.html, transactions.html (just the entry_type part), and added to dashbord.css.