Create of Lab documentation (#54)
* Creation of release report

This is the first draft of the progress release report. I followed the template and example posted on Blackboard in section 14. 

These are the following areas that need more detail: 
+ Testing and Quality Assurance
+ Bug Fixes and Enhancements
+ Deployment (confirm AWS deployment)
+ Bug Fixes
+ Known Issues
+ Additional Limitations (if you can think any)

* Creation of user manual

Here is the first rough draft of the user manual. I need to go back and add the images to accompany the step by step guides (per the template instructions). I think I will wait until the application is completely done before getting some of the images. Otherwise we'll have to change them.

The following sections need more detail: 
+ Notifications steps
+ Goals History (refine the steps)
+ Additional troubleshooting (if any)
+ Additional FAQ (if any)
+ Additional support & contact info

* Create installation_images

create installation images file

* Delete docs/installation_images

* Create pythonDownload

* Add files via upload

upload images for installation guide

* Create Installation Guide.md

Added images for installing python, pip, and django

* Rename instalPython.png to installPython.png

fixed type in image name. It should show up in the installation document

* Rename Installation Guide.md to docs/Installation Guide.md

moved the installation guide to the docs folder

* Update Installation Guide.md

fixed the path to the images after moving the guide to the docs folder.

* Update Installation Guide.md

added group info

* Add files via upload

upload additional images

* Update Installation Guide.md

Added additional pictures

* update user manual, format

---------

Co-authored-by: skybldev <jake@skybldev.eu.org>
Co-authored-by: Kenneth Stancil <145942639+familycrest@users.noreply.github.com>